Move geos-xml-viewer to geos-xml-tools
