+++ geos-xml-tools/src/geos/xml_tools/viewer/bin/exporter.py
@@ -4,10 +4,8 @@
 
 import argparse
 from pathlib import PurePath
-
-from geos_xml_viewer.filters.geosDeckReader import GeosDeckReader
-from vtkmodules.vtkIOParallelXML import (
-    vtkXMLPartitionedDataSetCollectionWriter, )
+from vtkmodules.vtkIOParallelXML import vtkXMLPartitionedDataSetCollectionWriter
+from geos.xml_tools.viewer.filters.geosDeckReader import GeosDeckReader
 
 
 def valid_file( param: str ) -> str:
