+++ geos-xml-tools/src/geos/xml_tools/viewer/bin/testCellLocator.py
@@ -1,5 +1,4 @@
 import argparse
-
 import pyvista as pv
 from vtkmodules.vtkCommonCore import (
     reference,
