+++ geos-xml-tools/tests/generate_test_xml.py
