--- geos-xml-viewer/pyproject.toml
@@ -1,101 +1,0 @@
-[build-system]
-requires = ["setuptools>=61.2"]
-build-backend = "setuptools.build_meta"
-
-[tool.setuptools]
-include-package-data = true
-
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["geos_xml_viewer*", "PVplugins*"]
-exclude = ['tests*']
-
-[project]
-name = "geos-xml-viewer"
-version = "1.2.1"
-description = "geos-xml-viewer is a Python package dedicated to preprocessing and postpressing of data for the geos simulation framework"
-authors = [{name = "Lionel Untereiner", email = "lionel.untereiner@external.totalenergies.com"}]
-license = {text = "Apache-2.0"}
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "Development Status :: 4 - Beta",
-    "Intended Audience :: Developers",
-    "Intended Audience :: Science/Research",
-    "License :: OSI Approved :: Apache Software License ",
-    "Natural Language :: English",
-    "Operating System :: POSIX :: Linux",
-    "Operating System :: MacOS :: MacOS X",
-    "Topic :: Scientific/Engineering :: Visualization",
-]
-keywords = [
-    "GEOS",
-    "Simulation",
-]
-dependencies = [
-	"pyvista >= 0.42",
-	"lxml >= 4.9",
-	"xsdata >= 24",
-	"colorcet >= 3.0.1",
-    "typing_extensions>=4.12",
-]
-requires-python = ">= 3.9"
-
-[project.urls]
-Homepage = "https://github.com/GEOS-DEV/geosPythonPackages"
-Documentation = "https://geosx-geosx.readthedocs-hosted.com/projects/geosx-geospythonpackages/en/latest/"
-Repository = "https://github.com/GEOS-DEV/geosPythonPackages.git"
-"Bug Tracker" = "https://github.com/GEOS-DEV/geosPythonPackages/issues"
-
-[project.optional-dependencies]
-cli = ["xsdata[cli]>=24"]
-build = [
-  "build >= 1.2"
-]
-dev = [
-    "yapf",
-    "mypy",
-    "xmlschema"
-]
-test = [
-    "pytest-cov", 
-    "pytest"
-]
-
-[project.scripts]
-geos-viewer = "geos_xml_viewer.bin.viewer:run"
-geos-exporter = "geos_xml_viewer.bin.exporter:run"
-#validate = "geos_xml_viewer.bin.validate:run"
-#xsd = "geos_xml_viewer.bin.test_xsdata:run"
-geos-modifier = "geos_xml_viewer.bin.modifier:run"
-#testCellLocator = "geos_xml_viewer.bin.testCellLocator:run"
-geos-splitter = "geos_xml_viewer.bin.splitter:run"
-
-[tool.bumpversion]
-current_version = "1.2.1"
-
-[[tool.bumpversion.files]]
-filename = "pyproject.toml"
-search = 'version = "{current_version}"'
-
-[tool.pytest.ini_options]
-addopts = [
-    "--import-mode=importlib",
-]
-console_output_style = "count"
-python_classes = "Test"
-python_files = "test_*.py"
-python_functions = "test*"
-testpaths = ["tests"]
-pythonpath = [
-    "src",
-]
-norecursedirs = "bin"
-filterwarnings = []
-
-[tool.coverage.run]
-branch = true
-source = ["geos"]
-omit = [
-    "*/PVplugins/*",
-]
-
