--- geos-xml-viewer/src/geos_xml_viewer/filters/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright 2023-2024 TotalEnergies.
-# SPDX-FileContributor: Lionel Untereiner
